+++ lib/e_menu/src/apave.tcl
@@ -7,7 +7,7 @@
 ###########################################################
 
 package require Tk
-package provide apave 4.6.6
+package provide apave 4.7.0
 
 source [file join [file dirname [info script]] apavedialog.tcl]