+++ lib/hl_tcl/tcl_html.tcl
@@ -24,3 +24,4 @@ foreach ghtml $::argv {
     }
   }
 }
+after idle exit